requirements.txt
allure-pytest
appium-python-client~=2.7.1
pytest
pytest-testconfig
requests
selenium
urllib3